+++ lib/karafka/params/params.rb
@@ -32,7 +32,7 @@ module Karafka
         def build(message, parser)
           # Hash case happens inside backends that interchange data
           if message.is_a?(Hash)
-            new(parser: parser).merge!(message)
+            new(parser: parser).send(:merge!, message)
           else
             # This happens inside Kafka::FetchedMessagesProcessor
             new(
@@ -57,11 +57,14 @@ module Karafka
       #   retrieve data, parser will be executed to parse data. Output of parsing will be merged
       #   to the current object. This object will be also marked as already parsed, so we won't
       #   parse it again.
-      def retrieve
+      def retrieve!
         return self if self[:parsed]
 
         merge!(parse(delete(:value)))
       end
+
+
+      private
 
       # Overwritten merge! method - it behaves differently for keys that are the same in our hash
       #  and in a other_hash - it will not replace keys that are the same in our hash
@@ -79,8 +82,6 @@ module Karafka
       def merge!(other_hash)
         super(other_hash) { |_key, base_value, _new_value| base_value }
       end
-
-      private
 
       # @param value [String] Raw data that we want to parse using controller's parser
       # @note If something goes wrong, it will return raw data in a hash with a message key